refactor: passwordConfirm을 userInfo로 관리 (#143)
