+++ src/components/SignUp/initialData.js
@@ -1,0 +1,34 @@
+const { default: ROLE } = require('constants/role.js');
+
+const initialUserInfo = {
+  id: '',
+  password: '',
+  passwordConfirm: '',
+  name: '',
+  cardInfo: {
+    cardNum: '',
+    expiredDate: '',
+    cvc: '',
+  },
+  address: '',
+  addressDetail: '',
+  age: '',
+  role: ROLE.TEACHER,
+};
+
+const initialError = {
+  id: '',
+  password: '',
+  passwordConfirm: '',
+  name: '',
+  cardInfo: {
+    cardNum: '',
+    expiredDate: '',
+    cvc: '',
+  },
+  address: '',
+  addressDetail: '',
+  age: '',
+};
+
+export { initialUserInfo, initialError };
